requirements.txt
requests>=2.25.1
openai==1.3.9
python-dotenv>=0.21.0
pandas